+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added category filter to location view
 - Added ability to archive an entire section of climbs
+- Added fullscreen view on image tap
 
 ## 0.3.0 - 2020-02-27